requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
alembic==1.13.1
pyaudio==0.2.14
SpeechRecognition==3.10.1
transformers==4.37.0
torch==2.1.2
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
